src/web/requirements.txt
Flask>=3.0
gunicorn>=22.0
requests>=2.31
azure-identity>=1.15
